Consolidate completion fns
