+++ src/subcommands/complete.rs
@@ -4,8 +4,9 @@ use std::io::{self, Write};
 use anyhow::Result;
 use structopt::{clap::AppSettings, StructOpt};
 
-use crate::config::{AppConfig, get};
-use crate::subcommands::{completion_containers, completion_images, completion_services, Subcommand};
+use crate::config::{AppConfig, ComposeFile, get};
+use crate::subcommands::{configured_yamls, get_yaml, Subcommand};
+use crate::term::external_output;
 use crate::util::*;
 
 #[derive(StructOpt)]
@@ -31,6 +32,30 @@ impl Subcommand for Complete {
             _ => Ok(()), // $ dock <subcommand> <arg> _  (empty return will invoke shell default completions)
         }
     }
+}
+
+fn completion_containers() -> Result<Vec<u8>> {
+    external_output("docker", crate::vec_of_strings!["ps", "--format", "{{.Names}}"])
+}
+
+fn completion_images(with_tags: bool) -> Result<Vec<u8>> {
+    match with_tags {
+        true => external_output("docker", crate::vec_of_strings!["images", "--format", "{{.Repository}}:{{.Tag}}"]),
+        _ => external_output("docker", crate::vec_of_strings!["images", "--format", "{{.Repository}}"]),
+    }
+}
+
+fn completion_services(mode: &String) -> Result<Vec<String>> {
+    let services = configured_yamls(mode)
+        .iter()
+        .map(|filename| get_yaml(filename).expect(filename))
+        .map(|ComposeFile { services }| services.keys()
+            .map(String::from)
+            .collect::<Vec<_>>()
+            .clone())
+        .flatten()
+        .collect::<Vec<_>>();
+    Ok(services.clone())
 }
 
 fn complete_subcommands(mode: &String) -> Result<()> {
