+++ src/subcommands/mod.rs
@@ -5,7 +5,7 @@ use walkdir::WalkDir;
 
 use crate::config::{AppConfig, ComposeFile, get};
 use crate::fs::read;
-use crate::term::{external_spawn, external_output};
+use crate::term::{external_spawn};
 use crate::util::*;
 
 pub mod alias;
@@ -52,30 +52,6 @@ pub mod all {
 
 pub trait Subcommand {
     fn process(&self, mode: Option<&String>) -> Result<()>;
-}
-
-fn completion_containers() -> Result<Vec<u8>> {
-    external_output("docker", crate::vec_of_strings!["ps", "--format", "{{.Names}}"])
-}
-
-fn completion_images(with_tags: bool) -> Result<Vec<u8>> {
-    match with_tags {
-        true => external_output("docker", crate::vec_of_strings!["images", "--format", "{{.Repository}}:{{.Tag}}"]),
-        _ => external_output("docker", crate::vec_of_strings!["images", "--format", "{{.Repository}}"]),
-    }
-}
-
-fn completion_services(mode: &String) -> Result<Vec<String>> {
-    let services = configured_yamls(mode)
-        .iter()
-        .map(|filename| get_yaml(filename).expect(filename))
-        .map(|ComposeFile { services }| services.keys()
-            .map(String::from)
-            .collect::<Vec<_>>()
-            .clone())
-        .flatten()
-        .collect::<Vec<_>>();
-    Ok(services.clone())
 }
 
 fn compose(args: Vec<String>, mode: &String) -> Result<()> {
